refactor(pdf-export): extract CharacterPdfExportButton state to factory
Moves isExporting state and handleExport function into
createCharacterPdfExportButtonState(), reducing the script tag from
59 lines to 9. Component now only holds imports, props, and factory call.
