+++ src/view/sheets/character/pdfExport/CharacterPdfExportButton.svelte
@@ -1,71 +1,23 @@
 <script lang="ts">
 	import type { CharacterPdfExportButtonProps } from '#types/components/CharacterPdfExportButton.js';
-
-	import GenericDialog from '#documents/dialogs/GenericDialog.svelte.ts';
 	import localize from '#utils/localize.ts';
-
 	import PdfExportDialog from '../../../dialogs/PdfExportDialog.svelte';
+	import { createCharacterPdfExportButtonState } from './CharacterPdfExportButtonState.svelte.ts';
 
 	const { actor }: CharacterPdfExportButtonProps = $props();
 
-	let isExporting = $state(false);
-
-	async function handleExport() {
-		if (isExporting) return;
-
-		// Open the PDF export dialog
-		const uniqueId = `pdf-export-${actor.id}`;
-
-		// Check if dialog is already open for this character
-		if (GenericDialog.isOpen(uniqueId)) {
-			return;
-		}
-
-		const dialog = GenericDialog.getOrCreate(
-			localize('NIMBLE.pdfExport.dialogTitle'),
-			PdfExportDialog as unknown as Parameters<typeof GenericDialog>[1],
-			{ actor },
-			{
-				icon: 'fa-solid fa-file-pdf',
-				width: 900,
-				height: Math.min(window.innerHeight - 100, 750),
-				uniqueId,
-			},
-		);
-
-		dialog.render(true);
-
-		// Wait for dialog result
-		const result = await dialog.promise;
-
-		if (result?.columnContent) {
-			isExporting = true;
-
-			try {
-				const { exportCharacterPdf } = await import('./exportCharacterPdf.ts');
-				await exportCharacterPdf(actor, {
-					columnContent: result.columnContent as [string, string, string],
-					template: result.template ?? 'lined',
-				});
-				ui.notifications?.info(localize('NIMBLE.pdfExport.success'));
-			} catch (error) {
-				console.error('PDF export failed:', error);
-				ui.notifications?.error(localize('NIMBLE.pdfExport.error'));
-			} finally {
-				isExporting = false;
-			}
-		}
-	}
+	const state = createCharacterPdfExportButtonState(() => actor, PdfExportDialog);
+	const { handleExport } = state;
 </script>
 
 <button
 	class="nimble-button"
 	data-button-variant="full-width"
 	type="button"
-	disabled={isExporting}
+	disabled={state.isExporting}
 	onclick={handleExport}
 >
-	{#if isExporting}
+	{#if state.isExporting}
 		<i class="fa-solid fa-spinner fa-spin"></i>
 		{localize('NIMBLE.pdfExport.exporting')}
 	{:else}
